refactor: create store
