+++ .env.example
@@ -1,0 +1,2 @@
+NODE_ENV='development'
+VERSION='1.0.0'
