+++ package.json
@@ -13,6 +13,7 @@
     "@hookform/resolvers": "^3.1.0",
     "@vercel/analytics": "^1.0.1",
     "date-fns": "^2.30.0",
+    "dotenv": "^16.1.1",
     "immer": "^10.0.2",
     "phosphor-react": "^1.4.1",
     "react": "^18.2.0",
@@ -24,6 +25,7 @@
   },
   "devDependencies": {
     "@rocketseat/eslint-config": "^1.2.0",
+    "@types/node": "^20.2.5",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "@types/react-router-dom": "^5.3.3",
