+++ src/store/reducers/cycles/actions.ts
@@ -1,10 +1,4 @@
-import { Cycle } from './reducer'
-
-export enum ActionTypes {
-  ADD_NEW_CYCLE = 'ADD_NEW_CYCLE',
-  INTERREPT_CURRENT_CYCLE = 'INTERREPT_CURRENT_CYCLE',
-  MARK_CURRENT_CYCLE_AS_FINISHED = 'MARK_CURRENT_CYCLE_AS_FINISHED',
-}
+import { ActionTypes, Cycle } from './types'
 
 export function addNewCycleAction(newCycle: Cycle) {
   return {
