+++ src/store/reducers/cycles/types.ts
@@ -1,0 +1,19 @@
+export enum ActionTypes {
+  ADD_NEW_CYCLE = 'ADD_NEW_CYCLE',
+  INTERREPT_CURRENT_CYCLE = 'INTERREPT_CURRENT_CYCLE',
+  MARK_CURRENT_CYCLE_AS_FINISHED = 'MARK_CURRENT_CYCLE_AS_FINISHED',
+}
+
+export interface Cycle {
+  id: string
+  task: string
+  minutesAmount: number
+  startDate: Date
+  interruptedDate?: Date
+  finishedDate?: Date
+}
+
+export interface CyclesState {
+  cycles: Cycle[]
+  activeCycleId: string | null
+}
